refactor(hooks): split pretty_output.py into private formatter modules
Splits hooks/pretty_output.py (781 lines) into the dispatch entrypoint plus
four underscore-prefixed private helper modules to bring it under the
350-line file budget (audit finding 8.3):

- _fmt_primitives.py: payload dataclasses, token formatter, pipeline-mode
  detector, short-name extractor, _HOOK_CONFIG_PATH_COMPONENTS
- _fmt_execution.py: run_skill, run_cmd, test_check, merge_worktree
- _fmt_status.py: token/timing summary, kitchen_status, clone_repo
- _fmt_recipe.py: load_recipe, open_kitchen, list_recipes

The dispatch entrypoint (pretty_output.py) re-imports the moved formatter
functions so the public surface (main, _format_response) is preserved.
Tests T-4 (file size budget) and T-5 (public surface guard) added.

Widens the hooks/ private-helper convention in arch tests so files starting
with underscore are recognised as imported helpers, not standalone hook
scripts: bumps hooks/ subpackage exemption from 14 to 18 files and skips
underscore-prefixed files in hook registration coverage checks.

Updates CLAUDE.md architecture map to document the four new private
formatter modules.
