+++ CLAUDE.md
@@ -254,7 +254,11 @@ generic_automation_mcp/
 │   ├── unsafe_install_guard.py
 │   ├── generated_file_write_guard.py
 │   ├── headless_orchestration_guard.py
-│   ├── pretty_output.py     #   Reformats MCP JSON responses as Markdown-KV
+│   ├── pretty_output.py     #   Dispatch entrypoint for MCP JSON → Markdown-KV reformatter
+│   ├── _fmt_primitives.py   #   Payload dataclasses, token formatter, pipeline-mode + short-name
+│   ├── _fmt_execution.py    #   run_skill, run_cmd, test_check, merge_worktree formatters
+│   ├── _fmt_status.py       #   token/timing summary, kitchen_status, clone_repo formatters
+│   ├── _fmt_recipe.py       #   load_recipe, open_kitchen, list_recipes formatters
 │   ├── token_summary_appender.py #  Appends Token Usage Summary to PR body
 │   └── session_start_reminder.py #  Injects open-kitchen reminder on resume
 │
